Update TSConfig to produce type definitions (#79)
* Update TSConfig to produce type definitions

Also, add some stricter type-checking, and remove settings that were set
to the same value as the default for that setting.

I also reorganized the sections into the sections that TypeScript itself
generates when you boot up a `tsconfig.json` using their CLI.

* Downgrade target to ES6 for Node 8 support

* Add engines specification to package.json

Make it clear that we support Node 8 and above.